Added a function to upload a file to the RMS (with some limitations).
The main limitation is that not all filenames are supported, due to
limitations in cURL. Double-quotes cannot be used, and most Unicode
characters will get mangled, so the function warns about the latter
and errors out on the former. It also warns about some risky ASCII
characters, that could potentially be misunderstood by the server.