Improved the rms_upload function to support most filenames.
This implements RFC5987 encoding of the filename and uploads as a raw
file instead of using cURL's form support. This removes the previous
limitations on the filename, but adds a new limitation of needing its
charset to be in UTF-8 or being convertable to that with iconv.

The old method is still available via a new parameter, to have the
widest possible support. (Also useful for debugging the RMS.)